% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{ping_nlp_client}
\alias{ping_nlp_client}
\title{Ping Stanford CoreNLP Server}
\usage{
ping_nlp_client(port = 9000L, host = "localhost",
  user_agent = "NLPclient")
}
\arguments{
\item{port}{an integer giving the port (default is \code{9000L}).}

\item{host}{a character string giving the hostname of the server.}

\item{user_agent}{a character string giving the client name.}
}
\value{
Returns \code{"pong"} if the server is online \code{""} otherwise.
}
\description{
Ping Stanford CoreNLP server.
}
\examples{
ping_nlp_client()
}
